refactor structure for agent builder
